feat(list_utils.c, push_swap.c, push_swap.h): add function to reverse linked list
Add a new function `reverseLinkedList` to `list_utils.c` that reverses a linked list. This function is used in `push_swap.c` to reverse the linked list after sorting. Update the `insertion_sort` function in `push_swap.c` to use a while loop instead of a for loop. Update the header file `push_swap.h` to include the new function declaration.